Refactor eye stream handling and fix data retrieval bug
This commit refactors the code related to handling the eye stream and fixes a bug in data retrieval. The eye stream was previously hardcoded, but now it is retrieved from the all_eye_data array. Additionally, a bug was fixed where data retrieval was not working properly when restreaming multiple streams in parallel. This bug was due to the while loop condition not being met. The code has been modified to ensure that exactly 108 samples are pulled from the eeg_inlet.

The commit also includes a commented out path variable that was used for testing purposes. This variable has been left for reference but is not used in the current implementation.

Lastly, the commit updates the computation of eye features. Previously, the eye features were set to an array of zeros, but now they are computed using the eye data.

No user commits were copied in this commit message.